+++ src/azure-cli/azure/cli/command_modules/storage/_client_factory.py
@@ -159,5 +159,9 @@ def cf_mgmt_blob_services(cli_ctx, _):
     return storage_client_factory(cli_ctx).blob_services
 
 
+def cf_mgmt_file_shares(cli_ctx, _):
+    return storage_client_factory(cli_ctx).file_shares
+
+
 def cf_blob_data_gen_update(cli_ctx, kwargs):
     return blob_data_service_factory(cli_ctx, kwargs.copy())