+++ src/azure-cli/HISTORY.rst
@@ -24,6 +24,10 @@ Release History
 
 * Fix #6371: Support filename and environment variable completion in Bash
 
+**Storage**
+
+* Add a new command group `az storage share-rm` to use the Microsoft.Storage resource provider for Azure file share management operations.
+
 2.0.80
 ++++++